feat!: allow arbitrary callback urls [UPL-62] (#310)
This removes the config option `callbackUrl` in favor of `Request.url`, to allow any callbackUrl instead of just `/api/uploadthing`.

- [x] docs

Should probably still be released as part of a major release, but it no longer requires the user to pass a url. We default to (process.env.VERCEL_URL / import.meta.env.VERCEL_URL / window.location.origin)/api/uploadthing.

Made some changes so that internally we always pass a URL object, asserting it's a valid absolute url. All internal functions should also require a URL be passed, but this required'ness should not be exposed to users

User can pass either just an origin, just a pathname, or a full url and we'll fallback to the closest one matching the default